ui2: port tr2 graphic settings dialog
