+++ src/tr2/game/option/option_detail.c
@@ -1,45 +1,57 @@
 #include "game/input.h"
 #include "game/option/option.h"
 #include "game/text.h"
-#include "game/ui/widgets/graphics_dialog.h"
+#include "game/ui2/dialogs/graphic_settings.h"
 #include "global/vars.h"
 
-static struct {
-    UI_WIDGET *widget;
-} m_Priv = {};
+typedef struct {
+    struct {
+        bool is_ready;
+        UI2_GRAPHIC_SETTINGS_STATE state;
+    } ui;
+} M_PRIV;
 
-static void M_ConstructUI(void);
-static void M_DestroyUI(void);
+static M_PRIV m_Priv = {};
 
-static void M_ConstructUI(void)
+static void M_Init(M_PRIV *p);
+static void M_Shutdown(M_PRIV *p);
+
+static void M_Init(M_PRIV *const p)
 {
-    m_Priv.widget = UI_GraphicsDialog_Create();
+    p->ui.is_ready = true;
+    UI2_GraphicSettings_Init(&p->ui.state);
 }
 
-static void M_DestroyUI(void)
+static void M_Shutdown(M_PRIV *const p)
 {
-    if (m_Priv.widget != nullptr) {
-        m_Priv.widget->free(m_Priv.widget);
-        m_Priv.widget = nullptr;
+    if (p->ui.is_ready) {
+        p->ui.is_ready = false;
+        UI2_GraphicSettings_Free(&p->ui.state);
     }
 }
 
 void Option_Detail_Control(INVENTORY_ITEM *const item, const bool is_busy)
 {
-    if (m_Priv.widget == nullptr) {
-        M_ConstructUI();
+    M_PRIV *const p = &m_Priv;
+    if (is_busy) {
+        return;
     }
-    m_Priv.widget->control(m_Priv.widget);
+    if (!p->ui.is_ready) {
+        M_Init(p);
+    }
+    UI2_GraphicSettings_Control(&p->ui.state);
 }
 
 void Option_Detail_Draw(INVENTORY_ITEM *const item)
 {
-    if (m_Priv.widget != nullptr) {
-        m_Priv.widget->draw(m_Priv.widget);
+    M_PRIV *const p = &m_Priv;
+    if (p->ui.is_ready) {
+        UI2_GraphicSettings(&p->ui.state);
     }
 }
 
 void Option_Detail_Shutdown(void)
 {
-    M_DestroyUI();
+    M_PRIV *const p = &m_Priv;
+    M_Shutdown(p);
 }
