--- src/tr2/game/ui/widgets/graphics_dialog.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/ui/widgets/base.h>
-
-UI_WIDGET *UI_GraphicsDialog_Create(void);
